mrpt-poses:isInfType method to comply to standard virtual inheritance
Commit also includes:

- Move STL related classes to stl_containers_utils.h

cc @jolting, @jlblancoc
