+++ libs/base/include/mrpt/math/utils.h
@@ -48,79 +48,7 @@ namespace mrpt
 		 */
 		template<class T>
 		T absDiff(const T& lhs, const T& rhs) {
-  			return lhs>rhs ? lhs-rhs : rhs-lhs;
-		}
-
-		/**\brief Return a STL container in std::string form.
-		 *
-		 * \param[in] t Template STL container (e.g. vector)
-		 * \return String form of given STL container
-		 */
-		template<class T>
-		std::string getSTLContainerAsString(const T& t) {
-			using namespace std;
-			stringstream ss;
-			for (typename T::const_iterator it = t.begin(); it != t.end(); ++it) {
-				ss << *it << ", ";
-			}
-			return ss.str();
-		}
-		/**\brief Print the given vector t.
-		 *
-		 * \param[in] t Template vector
-		 */
-		template<class T>
-		void printSTLContainer(const T& t) {
-			using namespace std;
-			cout << getSTLContainerAsString(t) << endl;
-		}
-		/**\brief Print the given STL container of containers t.
-		 *
-		 * \param[in] t Template STL container (containing other containers)
-		 */
-		template<class T>
-		void printSTLContainerOfContainers(const T& t) {
-			using namespace std;
-
-			int i = 0;
-			for (typename T::const_iterator it = t.begin();
-					it  != t.end(); ++i, ++it) {
-				cout << "List " << i+1 << "/" << t.size() << endl << "\t";
-				printSTLContainer(*it);
-			}
-		}
-
-		/**\brief Return contents of map in a string representation
-		 *
-		 * \param[in] m Template map
-		 * \param[in] sep String that seperates visually each key and its value. Defaults to " => "
-		 * \return std::string representation of map
-		 * */
-		template<class T1, class T2>
-		std::string getMapAsString(const std::map<T1, T2>& m,
-				const std::string& sep=" => ") {
-			using namespace std;
-			stringstream ss("");
-
-			for (typename map<T1, T2>::const_iterator it = m.begin();
-					it != m.end();
-					++it) {
-
-				ss << it->first << " => " << it->second << endl;
-			}
-
-			return ss.str();
-		}
-
-		/**\brief Print the given map m
-		 *
-		 * \param[in] m Template map
-		 */
-		template<class T1, class T2>
-		void printMap(const std::map<T1, T2>& m) {
-			using namespace std;
-
-			cout << getMapAsString(m) << endl;
+  			return lhs>rhs ? lhs - rhs : rhs - lhs;
 		}
 
 		/** \addtogroup container_ops_grp
