+++ libs/base/include/mrpt/poses/CPosePDFGaussianInf.h
@@ -14,15 +14,6 @@
 
 namespace mrpt
 {
-
-	namespace traits
-	{
-		template<>
-		struct is_inf_type< ::mrpt::poses::CPosePDFGaussianInf > {
-			static const bool value = true;
-		};
-	}
-
 	namespace poses
 	{
 
@@ -84,7 +75,7 @@ namespace mrpt
 			void getMean(CPose2D &mean_pose) const MRPT_OVERRIDE {
 				mean_pose = mean;
 			}
-			bool isInfType() const { return mrpt::traits::is_inf_type<self_t>::value; }
+			bool isInfType() const { return true; }
 
 			/** Returns an estimate of the pose covariance matrix (3x3 cov matrix) and the mean, both at once.
 		   * \sa getMean */
